requirements.txt
# 核心框架
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.8
langgraph>=0.0.40
langsmith>=0.0.90

# OpenAI API
openai>=1.10.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# 向量数据库
faiss-cpu>=1.7.4

# 配置管理
python-dotenv>=1.0.0

# 日志和工具
logging>=0.4.9.6 